improve show msg while disabling srv

+++ functions/F_01_ENV_01_disable_srv.sh
@@ -6,14 +6,19 @@
 #-----------------------------------------------------------------------------------------
 # Stop and disable services
 #-----------------------------------------------------------------------------------------
+local disable_msg
 for disabled_service in ${disabled_services[@]}
 do
-  echo ""
-  echo "-----------------------------"
-  echo "Stopping service \"${disabled_service}\" ..."
-  echo "-----------------------------"
-  systemctl stop ${disabled_service}
-  systemctl disable ${disabled_service}
-  echo ""
-  echo ""
+  disable_msg="$(systemctl disable ${disabled_service})"
+  if [[ $? -eq 0 ]]; then
+    echo ""
+    echo "-----------------------------"
+    echo "Stopping service \"${disabled_service}\" ..."
+    echo "-----------------------------"
+    echo -e "${disable_msg}"
+    systemctl stop ${disabled_service}
+    echo ""
+    echo ""
+
+  fi
 done
